Update fastq handling/args and Coupled file VariantSequence output
For fastq handling, done to conserve storage space:
- Parses `fastq` or `fastq.gz` files directly. Does not write `fastq.gz`
 files to new `fastq` files.
- `ParsedFilteredFastq` files are not saved by default. See args
`keep_parsed_fastqs` and `only_parse_fastqs`.

For the `[_Coupled_]` files:
- Change `ComboCoupled_` to `AminoAcidCoupled_` to keep nomenclature
consistent.
- The `VariantSequence` column now contains sequences related to the
`VariableRegion` from the `RefSeq` input, rather than the whole
contructed `RefSeq`. E.g., it removes the primer seed regions. For the
`paired_bp` sequences, it confirms that the primer seed is in the original
sequence and raises a flag in the output if not, since this region
should not vary.

Other notes:
- `evSeq` output folder is now `evSeqOutput` to be consistent with other
output files.


Former-commit-id: 43c292eec69a3ef2dafc1ffcd2a2d18707e16c56